fix: addressed pr comments
