+++ plugins/field-time/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "blockly-plugin-field-time",
-  "version": "0.0.0",
-  "description": "A Blockly plugin.",
+  "name": "@blockly/field-time",
+  "version": "1.0.0",
+  "description": "A Blockly time field.",
   "scripts": {
     "audit:fix": "blockly-scripts auditFix",
     "build": "blockly-scripts build",
@@ -21,14 +21,14 @@
     "blockly-plugin",
     "field-time"
   ],
-  "homepage": "https://github.com/team-vibe-labs/blockly-samples/tree/master/plugins\\field-time#readme",
+  "homepage": "https://github.com/google/blockly-samples/tree/master/field-time#readme",
   "bugs": {
-    "url": "https://github.com/team-vibe-labs/blockly-samples/issues"
+    "url": "https://github.com/google/blockly-samples/issues"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/team-vibe-labs/blockly-samples.git",
-    "directory": "plugins\\field-time"
+    "url": "https://github.com/google/blockly-samples.git",
+    "directory": "plugins/field-time"
   },
   "license": "Apache-2.0",
   "directories": {
@@ -47,7 +47,10 @@
   "peerDependencies": {
     "blockly": "^9.1.0"
   },
-  "publishConfig": {},
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://wombat-dressing-room.appspot.com"
+  },
   "eslintConfig": {
     "extends": "@blockly/eslint-config"
   },
